add beginGame and endGame logic to backend

+++ uploader/recorder.js
@@ -13,7 +13,10 @@ exports.begin = () => {
 }
 
 exports.end = () => {
-  process.kill()
+  if(process) {
+  	process.kill()
+  }
+  process = null
   return gameID
 }
